Increase the n initial value, this stops it going negative during the transient.
Tweak T BCs to be dirichlet_o2 then set the outer boundary cell to T_t.
This fixes the problem of negative T in the outermost cell without
resorting to fixing the whole BC, which causes an offset.
Note this means using DIFF_C2 on the T derivative so that the forced cell
isn't used